==> 1_Get_submissions.ipynb <==
import praw
import arrow
import pprint
import pickle

reddit = praw.Reddit(client_id='client_id',  # put yours
                     client_secret='client_secret',  # put yours
                     user_agent='user agent',
                     username='username',  # put yours
                     password='password')  # put yours

sub = reddit.subreddit('rateme')

created_stamp = sub.created_utc
created_time = arrow.get(created_stamp)

## Exploration

start = arrow.get(2015, 12, 14).timestamp
end = arrow.get(2015, 12, 19).timestamp
results = list(sub.submissions(start=start, end=end))

len(results)

import pickle

pickle.dump(results, open('test.bin', 'wb'))

restored = pickle.load(open('test.bin', 'rb'))

restored[3].comments.list()[6].created_utc

pprint.pprint(vars(restored[3]))

print(post.title)
print(post.score)
print(post.id)
print(post.url)
print(post.permalink)
author = post.author
print(author.name)

post = results[15]

pprint.pprint(vars(post))

post.comments.list()

replacement_res = post.comments.replace_more()  # I don't need the return value

# for i, comment in enumerate(post.comments):
#     print(i, comment.body)

len(post.comments.list())

results[0].comments[0].body

list(results[0].comments)[0].name

## Fetching submissions with comments to a large list and pickling it

results = []

cur_time = created_time
prev_year, prev_month = created_time.year, created_time.month

while cur_time < arrow.now():
    cur_end_time = cur_time.replace(days=0.5)
    cur_list = list(sub.submissions(start=cur_time.timestamp, end=cur_end_time.timestamp))
    results.extend(cur_list)
    if cur_time.month > prev_month or cur_time.year > prev_year:
        prev_year = cur_time.year
        prev_month = cur_time.month
        print(cur_time.year, cur_time.month)
        print(len(results))
        pickle.dump(results, open('rateme_submissions.dat', 'wb'))
    cur_time = cur_end_time
