==> 2_Extract_dataframes.ipynb <==
import pickle
import pandas as pd
import numpy as np
import arrow
import pprint
from collections import Counter
from datetime import datetime
from tqdm import tqdm

posts = pickle.load(open('rateme_submissions.dat', 'rb'))
print(len(posts))

# post = posts[0]
# comment = post.comments[0]

# for key in vars(post).keys():
#     print(vars(post)[key])

# posts_df = pd.DataFrame(posts)

posts_vars = [vars(post) for post in posts]

## Posts exploration

lists_of_keys = [list(post_vars.keys()) for post_vars in posts_vars]

keys = sorted(list(set([item for sublist in lists_of_keys for item in sublist])))

key_values = {key : [post_vars[key] if key in post_vars
                     else None for post_vars in posts_vars] for key in keys}

list(enumerate(keys))

Counter(key_values[keys[83]]).most_common()

medias = list(filter(lambda x: x is not None, key_values[keys[41]]))

indices = [i for i, x in enumerate(key_values[keys[80]]) if x is not None]

print(indices[:5])

# [x for x in key_values['media'] if x is not None and
#                                    'oembed' in x.keys() and
#                                    'url' in x['oembed'].keys() and
#                                    'youtube' in x['oembed']['url'].lower()]

# [post.title for post in [posts[index] for index in indices]]

comments - save in a separate df, not in vars, call it separately
author - needed
author_flair_text - None or not
created_utc - needed
distinguished - check and exclude posts with 'moderator'. others are None and needed
edited - False or not (others are True and timestamp)
gilded - 0 or 1
id
is_self - bool
link_flair_text - check and exclude if not None
media - if not None, extract fields oembed -> width, height, thumbnail_width, thumbnail_height.
    also can exclude youtube
over_18 - bool
permalink - needed to be able to go there manually
score - int
selftext
title

---
preview - one preview image if exists and links to different resolution versions of it

## Create DF

df = pd.DataFrame()

post = posts[1000]
row = {}

if post.distinguished == 'moderator':
#     break
    pass
if post.link_flair_text is not None:
    # break
    pass

row['author'] = post.author.name
row['author_has_flair'] = (post.author_flair_text is not None)
row['created'] = pd.Timestamp(datetime.utcfromtimestamp(post.created_utc))
row['edited'] = (post.edited != False)
row['gilded'] = (post.gilded == 1)
row['id'] = post.id
row['is_self'] = post.is_self
row['over_18'] = post.over_18
row['permalink'] = post.permalink
row['score'] = post.score
row['selftext'] = post.selftext
row['title'] = post.title

if post.media is not None and 'oembed' in post.media.keys():
    row['width'] = post.media['oembed']['width']
    row['height'] = post.media['oembed']['height']
    row['thumbnail_width'] = post.media['oembed']['thumbnail_width']
    row['thumbnail_height'] = post.media['oembed']['thumbnail_height']
else:
    row['width'] = row['height'] = row['thumbnail_width'] = row['thumbnail_height'] = None

# row

df = df.append([row])

df = pd.DataFrame()
for post in tqdm(posts):
    row = {}
    if post.distinguished == 'moderator':
        continue
    if post.link_flair_text is not None:
        continue
        
    row['author'] = post.author.name
    row['author_has_flair'] = (post.author_flair_text is not None)
    row['created'] = pd.to_datetime(datetime.utcfromtimestamp(post.created_utc))
    row['edited'] = (post.edited != False)
    row['gilded'] = (post.gilded == 1)
    row['id'] = post.id
    row['is_self'] = post.is_self
    row['over_18'] = post.over_18
    row['permalink'] = post.permalink
    row['score'] = post.score
    row['selftext'] = post.selftext
    row['title'] = post.title

    if post.media is not None and 'oembed' in post.media.keys():
        if 'width' in post.media['oembed'].keys():
            row['width'] = post.media['oembed']['width']
        if 'height' in post.media['oembed'].keys():
            row['height'] = post.media['oembed']['height']
        if 'thumbnail_width' in post.media['oembed'].keys():
            row['thumbnail_width'] = post.media['oembed']['thumbnail_width']
        if 'thumbnail_height' in post.media['oembed'].keys():
            row['thumbnail_height'] = post.media['oembed']['thumbnail_height']
    else:
        row['width'] = row['height'] = row['thumbnail_width'] = row['thumbnail_height'] = None
        
    df = df.append([row])

## To/from CSV. Failed for some obscure reason. I'll pickle.

df.to_csv('df.csv', encoding='utf8', index=False)

df.head()

df.shape

df.dtypes

new_df.dtypes

new_df = pd.read_csv('df.csv')

df.shape

# new_df[new_df.created.isnull()]

df.to_csv('df.csv', encoding='utf8', index=False)

def date_parser(time_series):
    return pd.to_datetime(time_series, errors='coerce')

new_df = pd.read_csv('df.csv', encoding = 'utf8', index_col=None,
                    dtype = {'created': str, 'author_has_flair': bool}, parse_dates=['created'], date_parser=date_parser)

# new_df = new_df.dropna(subset=['created'])
# new_df.author_has_flair = new_df.author_has_flair.astype(bool)
# new_df.edited = new_df.edited.astype(bool) # messy!
# new_df.author_has_flair = new_df.author_has_flair.astype(bool)
# new_df.author_has_flair = new_df.author_has_flair.astype(bool)
# new_df.author_has_flair = new_df.author_has_flair.astype(bool)

new_df.shape

Counter(new_df.author_has_flair)

## Pickling

pickle.dump(df, open('posts_df.dat', 'wb'))

posts_df = pickle.load(open('posts_df.dat', 'rb'))

## Comments

post = posts[1000]

post.id

# df[df.id.str.contains(post.id)]

comments = post.comments.list()

# indices = np.random.randint(0, 10000, 100)
# for index in tqdm(indices):
#     post = posts[index]
#     comments = post.comments.list()
#     for comment in comments:
#         if comment.downs > 0:
#             print(comment.downs)

author, body, ups, id, permalink, created_utc, edited, gilded, parent_id (remove first 3 letters), is_submitter, replies

comment = comments[3]

comment.parent_id

df = pd.DataFrame()
for i, post in tqdm(enumerate(posts)):
    if not posts_df.id.str.contains(post.id).any():
        continue
    comments_forest = post.comments
    replacement_res = comments_forest.replace_more()  # I don't need the return value
    comments = comments_forest.list()
    for comment in comments:
        if comment.author is None:
            continue
        row = {}
        row['name'] = comment.author.name
        row['body'] = comment.body
        row['score'] = comment.ups
        row['id'] = comment.id
        row['permalink'] = comment.permalink
        row['created'] = pd.to_datetime(datetime.utcfromtimestamp(comment.created_utc))
        row['edited'] = comment.edited
        row['gilded'] = (comment.gilded == 1)
        row['parent_post_id'] = comment.parent_id[3:] if comment.parent_id[:3] == 't3_' else None
        row['parent_comment_id'] = comment.parent_id[3:] if comment.parent_id[:3] == 't1_' else None
        if comment.parent_id[:3] != 't3_' and comment.parent_id[:3] != 't1_':
            pprint.pprint(vars(comment))
#         row['parent_id'] = comment.parent_id
        row['is_submitter'] = comment.is_submitter
        row['replies'] = [comment.id for comment in comment.replies.list()]
        
        df = df.append([row])
    if i % 1000 == 0:
        pickle.dump(df, open('comments_df.dat', 'wb'))

first = pickle.load(open('comments_df_until_17000.dat', 'rb'))

second = pickle.load(open('comments_df_17000_22000.dat', 'rb'))

third = pickle.load(open('comments_df_22000plus.dat', 'rb'))

full = pd.concat([first, second, third])

pickle.dump(full, open('comments_df.dat', 'wb'))
