==> 4_Extract_age_gender.ipynb <==
import numpy as np
import pandas as pd
from collections import Counter
import pickle
import string
import matplotlib.pyplot as plt
%matplotlib inline
import re

f_regex = re.compile('([^a-z]|^|$)f([^a-z]|^|$)')
m_regex = re.compile('([^a-z]|^|$)m([^a-z]|^|$)')
# num_regex = re.compile('[-+]?[.]?[\d]+(?:,\d\d\d)*[\.]?\d*(?:[eE][-+]?\d+)?')
num_regex = re.compile('[0-9]+')

def extract_gender(title):
    """returns a tuple: (gender, index of the title that contributed most to the decision)"""
    f_matches = [m.start()+1 for m in f_regex.finditer(title)]
    m_matches = [m.start()+1 for m in m_regex.finditer(title)]
    if (len(f_matches) > 0 and len(m_matches) > 0) or len(m_matches) > 1 or len(f_matches) > 1:
        return ('unicorn', 0)
    if len(f_matches) > 0:
        return ('female', f_matches[0])
    if len(m_matches) > 0:
        return ('male', m_matches[0])
    if title.find('female') != -1:
        return ('female', title.find('female'))
    if title.find('male') != -1:
        return ('male', title.find('male'))
    return ('unknown', 0)

def extract_age(title, gender_res):
    gender, gender_pos = gender_res
    age_list = [[int(title[m.start() : m.end()]), m.start()] for m in num_regex.finditer(title)]
    age_min = 18
    age_max = 80
    filtered = [item for item in age_list if (item[0] >= age_min and item[0] <= age_max)]
    if not filtered:
#         age = 0
        age = np.nan
    else:
#         age = 0
        age = np.nan
        min_dist = -1
        for num, num_index in age_list:
            if num_index > gender_pos:
                cur_dist = num_index - gender_pos
            else:
                cur_dist = gender_pos - num_index + 1
            if min_dist == -1 or cur_dist < min_dist:
                age = num
                min_dist = cur_dist
    return age

posts = pickle.load(open('data/posts_df.dat', 'rb'))
posts.index = posts.created
posts_b = posts.copy()

posts.title = posts['title'].str.lower()
posts.title = posts['title'].str.replace('\'', '')

gender_res = posts['title'].apply(lambda x: extract_gender(x))
posts['gender'] = gender_res.apply(lambda x: x[0])
posts = posts[posts.gender != 'unicorn']

posts['age'] = posts.apply(lambda row: extract_age(row['title'], extract_gender(row['title'])), axis=1)
# posts.age = posts.apply(lambda row: extract_age(row['title'], gender_res.loc[row.created]), axis=1)

### check manually that age is below 100 at this point

# # debug
# data = new_posts.sample(100).title.values
# for title in data:
#     gender_res = extract_gender(title)
#     extract_age(title, gender_res)

posts.sample(100)

pickle.dump(posts, open('data/posts_with_age_gender.dat', 'wb'))

# for title in data:
#     print(title)
#     gender_res = extract_gender(title)
#     print(gender_res)
#     print(extract_age(title, gender_res))
#     print()

# posts.resample('D').mean().is_self.plot(figsize=(20,10))

# posts.resample('D').mean().score.plot(figsize=(20,10))

# ser = pd.Series(1, index=dates)
# ser.resample('D').sum()

# print(posts[~posts.over_18].score.mean())
# print(posts[posts.over_18].score.mean())
