==> 5_Merge_ratings.ipynb <==
import pandas as pd
import numpy as np
import pickle
from tqdm import tqdm
from collections import Counter

## Compute mean, var, num of ratings for posts (obsolete)

posts = pickle.load(open('data/posts_with_age_gender.dat', 'rb'))
comments = pickle.load(open('data/comments_with_ratings_df.dat', 'rb'))

def noname():
    global ratings, posts
    for i, (index, row) in tqdm(enumerate(posts.iterrows())):
        ratings = comments[comments.parent_post_id == row.id].rating.values
        ratings = ratings[~np.isnan(ratings)]
        posts.set_value(index, 'ratings_mean', ratings.mean())
        posts.set_value(index, 'ratings_var', ratings.var())
        posts.set_value(index, 'ratings_num', ratings.shape[0])
        if i > 20:
            break

%load_ext line_profiler

%lprun -f noname noname()

for index, row in tqdm(posts.iterrows()):
    ratings = comments[comments.parent_post_id == row.id].rating.values
    ratings = ratings[~np.isnan(ratings)]
    posts.set_value(index, 'ratings_mean', ratings.mean())
    posts.set_value(index, 'ratings_var', ratings.var())
    posts.set_value(index, 'ratings_num', ratings.shape[0])

pickle.dump(posts, open('data/posts_with_ratings_summary.dat', 'wb'))

## Getting post ids of comments

comments = pickle.load(open('data/comments_with_ratings_df.dat', 'rb'))

# failures = 0
for i, (index, comment) in tqdm(enumerate(comments.iterrows())):
    try:  # to find root of the current comment tree and take its parent_post_id
        cur_comment = comment
        while cur_comment.parent_post_id is None:
#             print(cur_comment.parent_comment_id)
            cur_comment = comments[comments.id == cur_comment.parent_comment_id].iloc[0]
        comments.set_value(index, 'parent_post_id', cur_comment.parent_post_id)
    except:
        pass
#         failures += 1
#     if failures % 500 == 0:
#         print(str(failures) + '/' + str(i) + ' failures')

comments = comments[comments.parent_post_id.notnull()]

pickle.dump(comments, open('data/comments_with_post_ids.dat', 'wb'))

## pd.merge ratings and posts

comments = pickle.load(open('data/comments_with_post_ids.dat', 'rb'))

comments

posts_renamed = posts.rename(columns={col: 'post_' + col for col in posts.columns})
comments_renamed = comments.rename(columns={col: 'comment_' + col for col in comments.columns})

merged = posts_renamed.merge(comments_renamed, left_on='post_id',
                             right_on='comment_parent_post_id')

# pd.set_option("display.max_columns", 50)

merged = merged.drop(['comment_parent_post_id', 'comment_edited'], axis='columns')

merged = merged.rename(columns={'comment_name': 'comment_author'})

merged = merged[['post_gender', 'post_age', 'post_score', 'comment_rating', 'comment_score', 'post_author', 'comment_author',
        'post_title', 'comment_body', 'post_selftext',
        'post_created', 'comment_created', 'post_author_has_flair', 'post_edited', 'post_gilded',
        'post_height', 'post_width',
        'post_thumbnail_height', 'post_thumbnail_width',
        'post_is_self', 'post_over_18', 'comment_gilded', 'comment_is_submitter',
        'comment_body_len',
        'post_id', 'comment_id', 'comment_parent_comment_id', 'comment_replies',
        'post_permalink', 'comment_permalink']]

comment_author_counts = merged.comment_author.value_counts()
data['comment_author_num_comments'] = merged.comment_author.apply(lambda x: comment_author_counts.loc[x])

pickle.dump(merged, open('data/merged.dat', 'wb'))
